+++ src/commands/folders/createFolder.ts
@@ -2,28 +2,24 @@ import { Item, ItemGroup } from 'postman-collection';
 import { window } from 'vscode';
 import { Folder } from '../../postman';
 import { PostmanItemModel } from '../../views/postmanItems/postmanItemModel';
-import { getCollection } from '../../utils';
-import { runCommand } from '../commands';
+import { save } from '../../utils';
 
-export function createFolder(parentNode?: PostmanItemModel): void {
+export async function createFolder(parentNode?: PostmanItemModel): Promise<void> {
   if (parentNode === undefined || !(parentNode.isCollection() || parentNode.isFolder())) {
     return;
   }
 
-  window
-    .showInputBox({ placeHolder: 'Folder name' })
-    .then((name) => {
-      if (name === undefined || name === '') {
-        return;
-      }
+  const name = await window.showInputBox({ placeHolder: 'Folder name' });
+  if (name === undefined || name === '') {
+    return;
+  }
 
-      const parent = parentNode.itemObject;
+  const parent = parentNode.itemObject;
 
-      const itemGroup = new ItemGroup<Item>({ name });
-      const folder = new Folder(parent, itemGroup);
+  const itemGroup = new ItemGroup<Item>({ name });
+  const folder = new Folder(parent, itemGroup);
 
-      parent.addChild(folder);
+  parent.addChild(folder);
 
-      runCommand('saveCollection', getCollection(parent));
-    });
+  await save(parent);
 }